реорганизовать riddler: разбить крупные файлы по зонам ответственности
handler: quiz.go, question.go, list.go + хелпер toListItemResponse
repository: quiz.go (quiz_template), question.go (question + answer_option)
service: quiz.go, question.go
openapi: добавить теги Quizzes и Questions
