+++ riddler/internal/handler/quiz/quiz.go
@@ -1,0 +1,171 @@
+package quiz
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/google/uuid"
+
+	"riddler/internal/domain"
+	"riddler/internal/pkg/apperr"
+	"riddler/internal/pkg/httpx"
+	"riddler/internal/transport/dto"
+)
+
+func (h *Handler) CreateQuiz(c *gin.Context) {
+	userID, ok := userIDFromCtx(c)
+	if !ok {
+		return
+	}
+
+	var req dto.CreateQuizRequest
+	if err := c.ShouldBindJSON(&req); err != nil {
+		httpx.WriteError(c, apperr.ErrBadRequest)
+		return
+	}
+
+	var settings domain.QuizDefaultSettings
+	if req.DefaultSettings != nil {
+		settings = domain.QuizDefaultSettings{
+			TotalTimeLimitSec:    req.DefaultSettings.TotalTimeLimitSec,
+			QuestionTimeLimitSec: req.DefaultSettings.QuestionTimeLimitSec,
+		}
+	}
+
+	id, err := h.service.CreateQuiz(c.Request.Context(), userID, req.Title, req.Description, settings)
+	if err != nil {
+		httpx.WriteError(c, err)
+		return
+	}
+
+	c.JSON(http.StatusOK, dto.CreateQuizResponse{ID: id.String()})
+}
+
+func (h *Handler) GetQuiz(c *gin.Context) {
+	userID, ok := userIDFromCtx(c)
+	if !ok {
+		return
+	}
+
+	quizID, err := uuid.Parse(c.Param("id"))
+	if err != nil {
+		httpx.WriteError(c, apperr.ErrBadID)
+		return
+	}
+
+	detail, err := h.service.GetQuiz(c.Request.Context(), quizID, userID)
+	if err != nil {
+		httpx.WriteError(c, err)
+		return
+	}
+
+	questions := make([]dto.QuestionResponse, len(detail.Questions))
+	for i := range detail.Questions {
+		q := &detail.Questions[i]
+		options := make([]dto.AnswerOptionResponse, len(q.Options))
+		for j, o := range q.Options {
+			options[j] = dto.AnswerOptionResponse{
+				ID:        o.ID.String(),
+				Text:      o.Text,
+				IsCorrect: o.IsCorrect,
+			}
+		}
+		questions[i] = dto.QuestionResponse{
+			ID:         q.ID.String(),
+			Type:       string(q.Type),
+			Text:       q.Text,
+			ImageLink:  q.ImageLink,
+			OrderIndex: q.OrderIndex,
+			MaxScore:   q.MaxScore,
+			Metadata:   q.Metadata,
+			Options:    options,
+		}
+	}
+
+	c.JSON(http.StatusOK, dto.QuizDetailResponse{
+		ID:          detail.ID.String(),
+		Title:       detail.Title,
+		Description: detail.Description,
+		DefaultSettings: dto.QuizDefaultSettings{
+			TotalTimeLimitSec:    detail.DefaultSettings.TotalTimeLimitSec,
+			QuestionTimeLimitSec: detail.DefaultSettings.QuestionTimeLimitSec,
+		},
+		IsPublic:       detail.IsPublic,
+		IsDraft:        detail.IsDraft,
+		NeedEvaluation: detail.NeedEvaluation,
+		Questions:      questions,
+	})
+}
+
+func (h *Handler) UpdateQuiz(c *gin.Context) {
+	userID, ok := userIDFromCtx(c)
+	if !ok {
+		return
+	}
+
+	quizID, err := uuid.Parse(c.Param("id"))
+	if err != nil {
+		httpx.WriteError(c, apperr.ErrBadID)
+		return
+	}
+
+	var req dto.UpdateQuizRequest
+	if err := c.ShouldBindJSON(&req); err != nil {
+		httpx.WriteError(c, apperr.ErrBadRequest)
+		return
+	}
+
+	if err := h.service.UpdateQuiz(c.Request.Context(), quizID, userID, req.Title, req.Description); err != nil {
+		httpx.WriteError(c, err)
+		return
+	}
+
+	c.Status(http.StatusOK)
+}
+
+func (h *Handler) PublishQuiz(c *gin.Context) {
+	userID, ok := userIDFromCtx(c)
+	if !ok {
+		return
+	}
+
+	quizID, err := uuid.Parse(c.Param("id"))
+	if err != nil {
+		httpx.WriteError(c, apperr.ErrBadID)
+		return
+	}
+
+	var req dto.PublishQuizRequest
+	if err := c.ShouldBindJSON(&req); err != nil {
+		httpx.WriteError(c, apperr.ErrBadRequest)
+		return
+	}
+
+	if err := h.service.PublishQuiz(c.Request.Context(), quizID, userID, req.IsPublic); err != nil {
+		httpx.WriteError(c, err)
+		return
+	}
+
+	c.Status(http.StatusOK)
+}
+
+func (h *Handler) CopyQuiz(c *gin.Context) {
+	userID, ok := userIDFromCtx(c)
+	if !ok {
+		return
+	}
+
+	quizID, err := uuid.Parse(c.Param("id"))
+	if err != nil {
+		httpx.WriteError(c, apperr.ErrBadID)
+		return
+	}
+
+	newID, err := h.service.CopyQuiz(c.Request.Context(), quizID, userID)
+	if err != nil {
+		httpx.WriteError(c, err)
+		return
+	}
+
+	c.JSON(http.StatusOK, dto.CreateQuizResponse{ID: newID.String()})
+}
